Reinventing memory mechanism in my note service
Communication with the database has been implemented, mechanisms for synchronization and automatic periodic uploading of the cache to the database have also been implemented. The cache itself and a temporary cache of the latest changes associated with notes were implemented. Methods of communicating with this resulting structure. Part of the project has been rewritten to work with new methods.